Add 'new breadth was set' panel event,
use it instead of "before_collapse" and "before expand",

use it to delay the initial map.setCenter and the flying to a newly selected bldg.

Rename panel.expand to .resize_to_content. Because, in case of switching the content in an already opened panel, it's not actually expand, and in some cases it even shrinks the panel, if new content was smaller, so "expand" is slightly a misleading term.