Improve flyto-button
Disable button when the building's centroid is within map viewport (instead of "building is in the very center of viewport").
This simplifies thinking and code (the newly selected building is surely within viewport and so I don't need any special logic for initial button "disabled" state, it's just disabled for sure).

To detect building's position, analyze its centroid coords, instead of queryRenderedFeatures that proved to be unstable - if feature isn't yet rendered due to e.g. slow network, then check failed.

This coord analysis doesn't consider the panel,  only map viewport itself. This can be improved later.

Use throttle-with-trailing, instead of debounce, for update_flyto_button
+ make opacity transition faster;
the above 2 changes make the button feel more responsive

Insert it as "disabled" initially.
(It is inserted as part of try_open_building, and it means that such building is surely within the map viewport, and thus button surely needs to be disabled)

Increase distance btw "other" buttons in bldg details