Use drop-rate=1 argument for tippecanoe
Because it chose to drop some peaks at low Z.

Setting this arg fixed it. This arg means "keep all features"
As a result, overall tile folder size for Dalat went up by just about 10kb.

I believe there was similar problem with cable car points, and perhaps some fancy walkaround for this. So it's possible that now this walkaround can be removed