Add get_features method to tile_layers.
I think it has to replace tile_layer.feature_filter and _transform.
Though it's a bit strange that both renderables and tile_layers now have get_features(). Not sure it's a problem. But can be a symptom that they shall be one thing.

Add unesco_area title_layer in Hue

Draw a border for purple city, instead of fill (because it collides with unesco fill)