Rewrite  process_1_image script to js
The reason for this, 1st of all, is to be able to run this script without vurtual environment, because that's how I had to run it in order to have python dependencies available in wsl. JS script succeeds without any extra tricks

Write thumb and large image to a given city's folder. Given city is defined just by where the source file lives: it must be in cities_images/[city]/src/.....  Added the check that source path matches this pattern