On city load (i.e., on city change too), copy all from fids_to_img_names.mjs to city's dedicated localStorage item.
This will erase all that was previously in this ls item.
That is, unsaved stuff from ls will be killed irreversibly.
This is a bit sad but I took measures to ensure that changes are saved to file (fat warning is shown on the page), + it's a lot of code to make this super safe and tidy