Resolve() from try_fly_to_building even if bldg centroid wasn't found
This is necessary first of all to prevent one difficult debugging situation. If there are no centroids and  try_fly_to_building just "returns", it results in not executing a piece of code that expands the panel. So, clicking on a building with panel collapsed (and having no centroids for some reason), I saw a building selected but panel not expanded. Console shouted about absent centroids but it's easy to overlook due to much shit in console, + it' easy to miss the connection btw missing centroids and panel not being opened.
And in terms of ui, even if flying fails for some reason, it still makes total sense to expand the panel with newly selected building.