Fix renderables' style layers' selected variants not being generated in build_layers().
It's quite a rewrite in build_layers.

Set rules for drawing_importance: if it's not explicitly set on a layer, set importance to 0 (highest) for "symbol" layers and to Infinity for other.

Deep merge (instead of dumb replace) "layout" and "paint" props of "base" layer with same props of props_when_selected.

When merging filters with unknown amount of non-undefined members, make the resulting filter as shallow as possible, avoiding ["all", ...] stuff when unnecessary. Extract a "join_style_filter" util function for that.

In select_building, allow to set "id filter" in both "basic" and "composite" (["all", ...]) kind of filter