Don't warn about nonexistent geojson feature in case of a dead building
(In generate_features..., dead buildings geojson doesn't exist, because dead stuff is frontend-only now. Nevertheless, if dead building's id is present in fids_to_img_names, it will be checked for being present in tiled geojson. For this reason I had plenty of warning for dead buildings. So they are explicitly suppressed for dead. Unfortunately, similar problem may occur for any other frontend-only features, not only dead, but for now it's not the case)