Make titles of non-selectable buildings to be at "center" by default
Because there is no visual logic in placing them underneath (south), for there is no dark border in a non-selectable building, and the bare background doesn't obscure the title too much

In is_feature_selectable, use current_city from globalThis, instead of importing it from load_city. (It makes a lot of sense, because load_city is a frontend-only thing and it was wrongly imported into isomorphic thing)

Assign current_city to globalThis in load_city.

Make current_city a const, mutate it on load_city instead of reassigning it to another obj.

Make current_city object in generate_tiles.mjs
(not in build_city.mjs, because globalThis gets lost when calling a child script via execSync)