Introduce histoire,
to organize the knowledge of the currently selected id and previously selected id.

Start any navigation from pushing to histoire;
proceed to apply new id on the map and in the panel via subscription to histoire.

Make ids in url stand for not only building ids, but "menu" and other things.
The root (no-id) city url is "intro".

Introduce handle_id_change() that's called by subscription to histoire change. Inside it, decide on the action depending on what kind of content needs to be shown, and what was the cause of this change.

Make single-page links within a city work via data-href-id attribute on link elements: if generic body click handler detects that target has such attr, navigate to such id.

Try to restrict feature ids to numbers