Resize panel right after setting new content, if panel was expanded,
and not only when panel is going to actively expand.

This cures frozen breadth when navigating b/f to a building from a content of different size (another building or menu or any).

Navigation b/f to a building is a no-expand action, and thus resize was also skipped by asking the wrong question "should panel expand?", whereas it should resize for sure, if it was opened, and any new content is installed.